+++ src/modules/mouseHover/components/MouseHover.vue
@@ -18,6 +18,8 @@ export default {
             "titleFontFamily",
             "fontStyle",
             "titleFontStyle",
+            "fontWeight",
+            "titleFontWeight",
             "fontSize",
             "titleFontSize",
             "fontColor",
@@ -49,6 +51,7 @@ export default {
             :style="{
                 '--mousehover-font-family': fontFamily,
                 '--mousehover-font-style': fontStyle,
+                '--mousehover-font-weight': fontWeight,
                 '--mousehover-font-size': fontSize + 'px',
                 '--mousehover-font-color': fontColor,
                 '--mousehover-border-radius': infoBorderRadius + 'px',
@@ -71,6 +74,7 @@ export default {
                                 'font-family': titleFontFamily,
                                 'font-style': titleFontStyle,
                                 'font-size': titleFontSize + 'px',
+                                'font-weight': titleFontWeight,
                                 'color': titleFontColor
                             }"
                             v-html="text"
@@ -100,6 +104,7 @@ export default {
     font-size: var(--mousehover-font-size, $font-size-base);
     font-family: var(--mousehover-font-family, $font_family_default);
     font-style: var(--mousehover-font-style, normal);
+    font-weight: var(--mousehover-font-weight, normal);
     line-height: var(--mousehover-line-height);
     text-align: left;
     max-width: inherit;